+++ README.md
@@ -41,7 +41,4 @@
 	- 2. 입자들을 각 설정된 값에 따라 업데이트 -> 입자들의 배열을 UAV로 연결하여 CS 단계로 전달하고, CS단계에서 동일한 연산을 병렬 처리
 	- 3. 입자들의 배열을 렌더링 -> 업데이트된 입자들의 정보가 담긴 배열을 다시 CBV로 연결하여, 그래픽스 파이프라인에서 인스턴싱 기반 렌더링 처리
 
-
-- 계산 파이프라인을 추가하여, 파티클의 움직임 업데이트
-- 파티클이 의도한대로 움직이지는 않지만, 예제와 유사한 움직임 확인됨
--> Animate(), Render(0), Render(1) 순서에 따른 차이 발생 -> 문제 원인으로 예상됨
+SO 출력 버퍼의 값에서 오류 발생 중
